Unify Manifest generation
